+++ compute/src/worker.rs
@@ -20,8 +20,7 @@ use ekiden_core::error::{Error, Result};
 use ekiden_core::futures::sync::oneshot;
 use ekiden_core::futures::Future;
 use ekiden_core::rpc::api;
-use ekiden_storage_base::StorageBackend;
-use ekiden_storage_batch::BatchStorageBackend;
+use ekiden_storage_base::BatchStorage;
 use ekiden_untrusted::{Enclave, EnclaveContract, EnclaveDb, EnclaveIdentity, EnclaveRpc};
 
 use super::consensus::ConsensusFrontend;
@@ -57,14 +56,14 @@ struct WorkerInner {
     /// Contract running in an enclave.
     contract: Enclave,
     /// Storage backend.
-    storage: Arc<StorageBackend>,
+    storage: Arc<BatchStorage>,
     /// Enclave identity proof.
     #[allow(dead_code)]
     identity_proof: IdentityProof,
 }
 
 impl WorkerInner {
-    fn new(config: WorkerConfiguration, ias: Arc<IAS>, storage: Arc<StorageBackend>) -> Self {
+    fn new(config: WorkerConfiguration, ias: Arc<IAS>, storage: Arc<BatchStorage>) -> Self {
         measure_configure!(
             "contract_call_batch_size",
             "Contract call batch sizes.",
@@ -119,13 +118,14 @@ impl WorkerInner {
     ) -> Result<(OutputBatch, H256)> {
         measure_histogram!("contract_call_batch_size", batch.len());
 
-        // Prepare batch storage (perform up to 3 retries).
-        let batch_storage = Arc::new(BatchStorageBackend::new(self.storage.clone(), 3));
+        // Prepare batch storage.
+        self.storage.start_batch();
 
         // Run in storage context.
+        #[feature(coerce_unsized)]
         let (new_state_root, outputs) =
             self.contract
-                .with_storage(batch_storage.clone(), root_hash, || {
+                .with_storage(self.storage.clone(), root_hash, || {
                     measure_histogram_timer!("contract_call_batch_enclave_time");
 
                     // Execute batch.
@@ -135,7 +135,7 @@ impl WorkerInner {
         // Commit batch storage.
         {
             measure_histogram_timer!("contract_call_storage_commit_time");
-            batch_storage.commit().wait()?;
+            self.storage.end_batch().wait()?;
         }
 
         Ok((outputs?, new_state_root))
@@ -263,7 +263,7 @@ pub struct Worker {
 
 impl Worker {
     /// Create new contract worker.
-    pub fn new(config: WorkerConfiguration, ias: Arc<IAS>, storage: Arc<StorageBackend>) -> Self {
+    pub fn new(config: WorkerConfiguration, ias: Arc<IAS>, storage: Arc<BatchStorage>) -> Self {
         // Spawn inner worker in a separate thread.
         let (command_sender, command_receiver) = channel();
         thread::spawn(move || {