+++ scripts/test-e2e.sh
@@ -36,7 +36,7 @@ run_compute_node() {
         --compute-replicas 2 \
         --time-source-notifier system \
         --entity-ethereum-address 627306090abab3a6e1400e9345bc60c78a8bef57 \
-        --storage-backend remote \
+        --batch-storage immediate_remote \
         --port ${port} \
         --node-key-pair ${WORKDIR}/tests/committee_3_nodes/node${id}.key \
         --test-contract-id 0000000000000000000000000000000000000000000000000000000000000000 \
@@ -59,7 +59,7 @@ run_compute_node_storage_multilayer() {
         --compute-replicas 2 \
         --time-source-notifier system \
         --entity-ethereum-address 627306090abab3a6e1400e9345bc60c78a8bef57 \
-        --storage-backend multilayer \
+        --batch-storage multilayer \
         --storage-multilayer-sled-storage-base "$db_dir" \
         --storage-multilayer-aws-region us-west-2 \
         --storage-multilayer-aws-table-name test \